WORKOUTWITHAI/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pydantic==2.11.3
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
celery==5.5.1
redis==5.0.1
prometheus-client==0.20.0
pytest==8.3.5
boto3==1.34.72
opencv-python==4.9.0.80
mediapipe==0.10.21
numpy==1.26.4
huggingface-hub==0.23.0
openai==1.16.0
pydub==0.25.1
pillow==11.0.0
matplotlib==3.8.4
tqdm==4.66.4
